requirements.txt
streamlit>=1.28
langchain>=0.0.217
openai>=1.2
langchain_google_genai